requirements.txt
ansible-vault-win==1.0.0
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
distro==1.7.0
dj-database-url==1.0.0
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
envparse==0.2.0
oauthlib==3.2.2
packaging==21.3
paramiko==2.11.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
texttable==1.6.4
urllib3==1.26.12
django-filter~=22.1
pydantic~=1.10.2
parameterized~=0.8.1